Don't add/drop child columns when added/dropped to/from parent. Fixes #33.
 * pyrseas/dbobject/column.py (ColumnDict.diff_map): Don't drop if
   column has inherited attribute.
 * pyrseas/dbobject/table.py (Table.diff_map): Don't add column if it
   has an inherited attribute.  Some pep8-suggested changes elsewhere.
 * tests/dbobject/test_column.py: Two new tests to verify fix.

+++ pyrseas/dbobject/column.py
@@ -217,7 +217,8 @@ class ColumnDict(DbObjectDict):
             if (sch, tbl) in list(self.keys()):
                 for col in self[(sch, tbl)]:
                     if col.name not in [c.name for c in incols[(sch, tbl)]] \
-                            and not hasattr(col, 'dropped'):
+                            and not hasattr(col, 'dropped') \
+                            and not hasattr(col, 'inherited'):
                         stmts.append(col.drop())
 
         return stmts
